+++ lib/active_storage_validations/railtie.rb
@@ -2,15 +2,18 @@
 
 module ActiveStorageValidations
   class Railtie < ::Rails::Railtie
-    initializer "active_storage_validations.configure", after: :load_config_initializers do
+    # Using after: :load_config_initializers would cause a stack level too deep error
+    # See: https://github.com/igorkasyanchuk/active_storage_validations/issues/364
+
+    initializer "active_storage_validations.configure" do
       ActiveSupport.on_load(:active_record) do
-        send :include, ActiveStorageValidations
+        include ActiveStorageValidations
       end
     end
 
     initializer "active_storage_validations.extend_active_storage_blob" do
       ActiveSupport.on_load(:active_storage_blob) do
-        include(ActiveStorageValidations::ASVBlobMetadatable)
+        include ActiveStorageValidations::ASVBlobMetadatable
       end
     end
   end